requirements.txt
qrcode[pil]
Pillow>=9.0.0
